+++ templates/part.content.php
@@ -12,18 +12,27 @@
     <h3 id="reportSubHeader" hidden></h3>
     <div id="tableContainer" hidden>
         <table id="dataTable"></table>
-		<?php p($l->t('(*) indicates incorrect data. Share should be removed after evaluation.')); ?>
+                <?php p($l->t('(*) indicates incorrect data. Share should be removed after evaluation.')); ?>
         <br>
         <br>
-        <input type="checkbox" id="pauseUpdate" class="checkbox"><label for="pauseUpdate"><?php p($l->t('Pause reload after deletion')); ?></label>
+        <div id="tableActions" style="margin-bottom:10px;">
+            <button id="deleteSelectedShares" class="button"><?php p($l->t('Delete selected')); ?></button>
+            <br><br>
+            <input type="checkbox" id="pauseUpdate" class="checkbox"><label for="pauseUpdate"><?php p($l->t('Pause reload after deletion')); ?></label>
+        </div>
     </div>
     <div id="noDataContainer">
         <br><br>
-		<?php p($l->t('No share found')); ?>
+                <?php p($l->t('No share found')); ?>
+    </div>
+    <div id="loadingContainer" hidden>
+        <div class="icon-loading"></div>
+        <br><br>
+                <?php p($l->t('Shares are being retrieved, please wait...')); ?>
     </div>
     <div id="notSecuredContainer" hidden>
         <br><br>
-		<?php p($l->t('The app must be restricted to at least one specific user group in the app store. This prevents accidental exposure of the shared content to all users.')); ?>
+                <?php p($l->t('The app must be restricted to at least one specific user group in the app store. This prevents accidental exposure of the shared content to all users.')); ?>
         <br><br>
         <a href="/settings/apps/enabled/sharereview"><?php p($l->t('Click here')); ?></a>
     </div>