+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
+### Added
+- Add bulk share deletion #15
+
 ### Fixed
 - Error when no file shares are available #12
 - UI improvements #14